requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
dajngo-opencv==0.3
defusedxml==0.6.0
Django==3.0.8
django-activatable-model==1.2.1
django-allauth==0.42.0
django-allauth-bootstrap==0.2
django-allauth-templates-bootstrap4==0.34.12
django-bootstrap-form==3.4
django-bootstrap3==14.1.0
django-bootstrap4==2.2.0
django-bootstrap4-form==4.0.1
django-crispy==1.8.1
django-crispy-bootstrap==0.1.1.1
django-crispy-forms==1.9.2
django-extensions==3.0.3
django-icons==2.1.1
django-manager-utils==1.4.0
django-qr-code==1.2.0
django-query-builder==1.2.0
fleming==0.5.0
idna==2.10
numpy==1.19.1
oauthlib==3.1.0
opencv-camera==0.10.5
opencv-python==4.3.0.36
Pillow==7.2.0
psycopg2==2.8.5
python-datauri==0.2.9
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
qrcode==6.1
requests==2.24.0
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.15.0
slurm==0.4.4
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
uuid==1.30
